+++ codegen/allapigen.py
@@ -4,8 +4,8 @@ from time import time
 
 from ansys.fluent.core import CODEGEN_OUTDIR, FluentMode, FluentVersion, launch_fluent
 from ansys.fluent.core.codegen import StaticInfoType, allapigen, print_fluent_version
+from ansys.fluent.core.search import _search
 from ansys.fluent.core.utils.fluent_version import get_version_for_file_name
-from ansys.fluent.core.utils.search import _search
 
 if __name__ == "__main__":
     t0 = time()
