+++ src/ansys/fluent/core/__init__.py
@@ -29,12 +29,12 @@ from ansys.fluent.core.launcher.pyfluent_enums import (  # noqa: F401
     Precision,
     UIMode,
 )
+from ansys.fluent.core.parametric import LocalParametricStudy  # noqa: F401
+from ansys.fluent.core.search import search  # noqa: F401
 from ansys.fluent.core.services.batch_ops import BatchOps  # noqa: F401
 from ansys.fluent.core.session import BaseSession as Fluent  # noqa: F401
 from ansys.fluent.core.utils import fldoc
 from ansys.fluent.core.utils.fluent_version import FluentVersion  # noqa: F401
-from ansys.fluent.core.utils.parametric import LocalParametricStudy  # noqa: F401
-from ansys.fluent.core.utils.search import search  # noqa: F401
 from ansys.fluent.core.utils.setup_for_fluent import setup_for_fluent  # noqa: F401
 from ansys.fluent.core.warnings import (  # noqa: F401
     PyFluentDeprecationWarning,
