+++ tests/test_search.py
@@ -4,7 +4,7 @@ from util.meshing_workflow import new_watertight_workflow_session  # noqa: F401
 from util.solver_workflow import new_solver_session  # noqa: F401
 
 import ansys.fluent.core as pyfluent
-from ansys.fluent.core.utils.search import (
+from ansys.fluent.core.search import (
     _get_api_tree_data,
     _get_capitalize_match_for_word_from_names,
     _get_close_matches_for_word_from_names,
